tests/submission/Infraware/Forms/contents/Forms/output_labels.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>output_labels</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action="" 
		  id="input_form">
	 <p><label>Full name<label>(name):<output name=o1 id='output_id1'>John</output></label></label></p>
	 <p><label>Age:<output name=o2 id='output_id2'>26</output></label></p>
	</form>
	<script>

		var output1 = document.getElementById("output_id1");
		var output2 = document.getElementById("output_id2");

		try {
			if (typeof(output1.labels) == "object") {
				if (output1.labels.length == 2 && output2.labels.length == 1) {
					test(function() {
						assert_true(true, "labels attribute is correct.");
					});
				} else {
					test(function() {
						assert_unreached("labels attribute is not correct.");
					});
				}
			} else {
				test(function() {
					assert_unreached("labels attribute is not exist.");
				});
			}
		} catch(e) {
			test(function() {
					assert_unreached("Error is raised.");
				});
		}
			
	</script>

 </body>
</html>
